litrs::escape

Function unescape

pub(crate) fn unescape(
    input: &str,
    unicode: bool,
    byte_escapes: bool,
    allow_nul: bool,
) -> Result<(Unescape, usize), ParseError>

Must start with \. Returns the unscaped value as E and the number of input bytes the escape is long.

unicode and byte_escapes specify which types of escapes are supported. Quote escapes are always unescaped, Unicode escapes only if unicode is true. If byte_escapes is false, ASCII escapes are used, if it’s true, Byte escapes are (the only difference being that the latter supports \xHH escapes > 0x7f).