Trait litrs::FromIntegerLiteral

pub trait FromIntegerLiteral: Sealed + Copy { }

Integer literal types. Implementation detail.

Implemented for all integer literal types. This trait is sealed and cannot be implemented outside of this crate. The trait’s methods are implementation detail of this library and are not subject to semver.

Object Safety

This trait is not object safe.